Made adjustments MVVM
MVVM template adjusted
